Add server auto disconnect and health maintener
The new version of ther server can automatically shutdown a consumer and
send an event back yo him if he is alive. Therefore the consumer can
reconnect. And it adds the term of ttl to keep alive a consumer.

Those two features are taken into account in this new version